[#2] Added support for storing half stars in ratings.

+++ src/main/java/com/example/ratingbadge/model/Rating.java
@@ -22,5 +22,5 @@ public class Rating extends BaseUUID {
     @Column(length = 300)
     private String review;
 
-    private int stars;
+    private float stars;
 }
